+++ CHANGES.md
@@ -1,3 +1,6 @@
+## 0.1.32 - UNRELEASED
+
+
 ## 0.1.31 - 7 Jul 2015
 
 Updated Clojure dependency to 1.7.0.